Fix code review issues: improve type safety and null termination
- Use explicit DWORD64 casts for 64-bit arithmetic in file size calculation
- Use compile-time sizeof for string lengths instead of runtime strlen
- Fix buffer overflow protection to ensure space for null terminator

+++ trunk/pe32+/pe32+/pe32+.cpp
@@ -96,7 +96,7 @@ bool MapFileForReading(wchar_t * filename, FileMappingInfo & info) {
 
     DWORD FileSizeHigh;
     DWORD FileSizeLow = GetFileSize(info.hfile, &FileSizeHigh);
-    info.filesize = ((DWORD64)FileSizeHigh << 32) | FileSizeLow;
+    info.filesize = ((DWORD64)FileSizeHigh << 32) | (DWORD64)FileSizeLow;
 
     if (info.filesize == 0) {
         return false;
